category seeder create

+++ database/migrations/2021_10_12_000000_create_users_table.php
@@ -15,9 +15,9 @@ class CreateUsersTable extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('language_id')->nullable();
+            $table->unsignedBigInteger('language_id')->nullable();
             $table->foreign('language_id')->references('id')->on('languages');
-            $table->unsignedInteger('currency_id')->nullable();;
+            $table->unsignedBigInteger('currency_id')->nullable();;
             $table->foreign('currency_id')->references('id')->on('currencies');
             $table->string('name');
             $table->string('email')->unique();
